refactor: 예외 이름 변경
