--- src/main/java/wooteco/subway/exceptions/SubWayCustomException.java
@@ -1,19 +1,0 @@
-package wooteco.subway.exceptions;
-
-public class SubWayCustomException extends RuntimeException {
-
-    private final SubWayException subWayException;
-
-    public SubWayCustomException(SubWayException subWayException) {
-        super(subWayException.message());
-        this.subWayException = subWayException;
-    }
-
-    public String message() {
-        return subWayException.message();
-    }
-
-    public int status() {
-        return subWayException.status();
-    }
-}
