+++ src/main/java/wooteco/subway/exceptions/SubWayException.java
@@ -1,27 +1,19 @@
 package wooteco.subway.exceptions;
 
-import org.springframework.http.HttpStatus;
+public class SubWayException extends RuntimeException {
 
-public enum SubWayException {
+    private final SubWayExceptionSet subWayExceptionSet;
 
-    DUPLICATE_EMAIL_EXCEPTION("중복된 이메일 입니다.", HttpStatus.BAD_REQUEST.value()),
-    NOT_EXIST_STATION_EXCEPTION("존재하지 않는 역입니다.", HttpStatus.BAD_REQUEST.value()),
-    NOT_CONNECT_STATION_EXCEPTION("연결되지 않은 역입니다.", HttpStatus.BAD_REQUEST.value());
-
-
-    private final String message;
-    private final int status;
-
-    SubWayException(String message, int status) {
-        this.message = message;
-        this.status = status;
+    public SubWayException(SubWayExceptionSet subWayExceptionSet) {
+        super(subWayExceptionSet.message());
+        this.subWayExceptionSet = subWayExceptionSet;
     }
 
     public String message() {
-        return message;
+        return subWayExceptionSet.message();
     }
 
     public int status() {
-        return status;
+        return subWayExceptionSet.status();
     }
 }
