+++ src/main/java/wooteco/subway/path/application/DijkstraAlgorithm.java
@@ -4,8 +4,8 @@ import java.util.List;
 import org.jgrapht.alg.shortestpath.DijkstraShortestPath;
 import org.jgrapht.graph.DefaultWeightedEdge;
 import org.jgrapht.graph.WeightedMultigraph;
-import wooteco.subway.exceptions.SubWayCustomException;
 import wooteco.subway.exceptions.SubWayException;
+import wooteco.subway.exceptions.SubWayExceptionSet;
 import wooteco.subway.path.dto.PathDto;
 import wooteco.subway.path.dto.PathRequest;
 
@@ -24,7 +24,7 @@ public class DijkstraAlgorithm implements PathAlgorithms {
             int distance = (int) dijkstraShortestPath.getPath(sourceId, targetId).getWeight();
             return new PathDto(paths, distance);
         } catch (NullPointerException e) {
-            throw new SubWayCustomException(SubWayException.NOT_CONNECT_STATION_EXCEPTION);
+            throw new SubWayException(SubWayExceptionSet.NOT_CONNECT_STATION_EXCEPTION);
         }
     }
 }
